pyproject.toml
[tool.poetry]
name = "autogpt"
version = "0.1.0"
description = "An AutoGPT agent."
authors = ["Tom Rochette <tom@tomrochette.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
openai = "^0.27.4"
python-dotenv = "^1.0.0"
tiktoken = "^0.3.3"
dask = "^2023.4.0"
distributed = "^2023.4.0"
bokeh = "^3.1.0"
prometheus-client = "^0.16.0"
tortoise-orm = "^0.19.3"
structlog = "^23.1.0"
redis = {extras = ["hiredis"], version = "^4.5.4"}
notion-client = "^2.0.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.21.0"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
